Split configuration loading to a separate file in todoapi command
