+++ cmd/todoapi/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"os"
-	"strings"
 
 	_ "github.com/lib/pq"
 	"github.com/overkilling/overkill-todo-monolith-api/http"
@@ -10,20 +9,6 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/spf13/viper"
 )
-
-func configureViper() {
-	viper.SetDefault("db.host", "localhost")
-	viper.SetDefault("db.port", 5432)
-	viper.SetDefault("db.database", "todo")
-	viper.SetDefault("db.username", "postgres")
-	viper.SetDefault("db.password", "postgres")
-
-	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
-	viper.AutomaticEnv()
-	viper.AddConfigPath(".")
-
-	viper.ReadInConfig()
-}
 
 func main() {
 	configureViper()
